pyproject.toml
[tool.poetry]
name = "ogu-api"
version = "0.1.0"
description = ""
authors = ["CraazzzyyFoxx <andrey.pod19@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.92.0"
tortoise-orm = "^0.19.3"
asyncpg = "^0.27.0"
uvicorn = "^0.20.0"
cashews = "^5.3.0"
httpx = "^0.23.3"
orjson = "^3.8.7"
pytz = "^2022.7.1"
selenium = "^4.8.2"
fake-useragent = "^1.1.1"
loguru = "^0.6.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
